finished project

+++ shapes/src/main/kotlin/Questions/Question.kt
@@ -4,6 +4,8 @@ import Shapes.Shape
 
 abstract class Question(val shape: Shape) {
     abstract val questionText: String
-    val answer: String = shape.getArea().toString()
+
+    //round answer to nearest integer
+    val answer: String = shape.getArea().toInt().toString()
 
 }
